Reorganize project and add googletest support
