+++ .circleci/config.yml
@@ -31,7 +31,7 @@ jobs:
           command: "cd ./build;cmake --build ./"
       - run:
            name: "Run Tests"
-           command: "cd ./build;ctest"
+           command: "cd ./build;ctest -j10 -C Debug -T test --output-on-failure"
 
 # Invoke jobs via workflows
 # See: https://circleci.com/docs/2.0/configuration-reference/#workflows
