+++ .gitignore
@@ -39,3 +39,4 @@
 /build
 .vscode
 /docs
+
