+++ src/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.23.2)
+
+add_executable(PLCompiler Main.cpp)
+
+add_subdirectory(libs)
+target_link_libraries(PLCompiler PRIVATE Lexer)
