--- src/Lexer/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_library(Lexer Lexer.cpp Token.cpp Tokenizer.cpp)
