+++ src/libs/Lexer/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(Lexer 
+    STATIC 
+    Lexer.cpp 
+    Token.cpp
+    Tokenizer.cpp
+)
