--- tests/TestPrograms/singleTokens.pl
@@ -1,2 +1,0 @@
-#keywords#
-if
